S15.e.1: extract operator.ts types + pure helpers (2894 → 2739 LOC)
First sub-slice of S15.e (operator.ts decomposition multi-PR sub-train).

Lifts the only purely-declarative + closure-free pieces of operator.ts:

- Module-level types (HealthState, SandboxInfo, EgressDomain,
  SecurityState, NodeInfo, ClusterHealth, MeshHealth) →
  cli/src/commands/operator/types.ts (128 LOC).
- Module-level pure helpers (timeSince, sumPrometheusCounter) →
  cli/src/commands/operator/helpers.ts (65 LOC).

operator.ts imports both modules; bodies and declarations are
byte-identical (no shape edits, no logic edits). Continues the
operator/ subdirectory pattern already started by operator/keymap.ts.

Verification:
- operator.ts LOC: 2894 → 2739
- tsc --noEmit / lint (27 warnings, baseline) / build: clean
- npm test: 454 pass / 2 skipped pre-existing

S15.e cap (800) not yet met — this slice sets scaffolding for
follow-up sub-slices: fetchers/, render/, actions/, dialogs/. Each
sub-slice ships independently behind its own audit doc.

Audit: docs/security-audits/2026-04-29-phase2-hotspot-operator-cli-e1.md
